requirements.txt
Flask==2.3.3
supabase==1.0.3
python-dotenv==1.0.0
pyjwt==2.8.0
python-decouple==3.8